+++ docs/api/index.rst
@@ -10,3 +10,4 @@ Workflows
 .. automodule:: fmriprep.workflows.base
 .. automodule:: fmriprep.workflows.anatomical
 .. automodule:: fmriprep.workflows.bold
+.. automodule:: fmriprep.workflows.fieldmap
