--- docs/sdc/base.rst
@@ -1,11 +1,0 @@
-:orphan:
-
-.. _sdc-base:
-
-Identifying field-mapping resources
------------------------------------
-
-.. automodule:: fmriprep.workflows.fieldmap.base
-    :members:
-    :undoc-members:
-    :show-inheritance:
