--- docs/sdc/unwarping.rst
@@ -1,11 +1,0 @@
-:orphan:
-
-.. _sdc-unwarp:
-
-Unwarping EPIs
---------------
-
-.. automodule:: fmriprep.workflows.fieldmap.unwarp
-    :members:
-    :undoc-members:
-    :show-inheritance:
